Extract rules UI component
